test: move horizontal scaling unit coverage
